WP8: Use cookies directly from the webview
In the previous version the cookies from the webview were copied to the
webrequest. When getting the cookies from the webiew only the cookies
for the current URL can be retrieved. This means that cookies of other
domains were never copied to the webrequest for upload and download.

Fixed by using the BrowserHttp implementation of IWebRequestCreate
which shares the cookie store with the webview. See also
http://stackoverflow.com/questions/4212713/grabbing-cookies-in-web-browser-control-wp7

However as a consequence if a cookie header is passed it will be
ignored.

Co-Authored-By: Leroy van Engelen <leroy.van.engelen@mendix.com>